第八题

+++ src/suyeq/Test.java
@@ -9,7 +9,8 @@ package suyeq;
  */
 public class Test {
     public static void main(String[] args){
-        System.out.println(Integer.MAX_VALUE);
+        System.out.println(Long.MAX_VALUE);
         System.out.println(Integer.MIN_VALUE);
+
     }
 }
